+++ multibyte.c
@@ -69,7 +69,7 @@ int
 pg_CS_code(const UCHAR *characterset_string)
 {
 	int i, c = -1;
-  	unsigned len = 0;
+  	size_t len = 0;
 
 	for(i = 0; CS_Table[i].code != OTHER; i++)
 	{
